refactor: 작업 설정 및 클라이언트 클래스 이름 리팩터링
