--- zipsoonbatch/src/main/java/com/zipsoon/zipsoonbatch/job/TestJobConfig.java
@@ -1,35 +1,0 @@
-package com.zipsoon.zipsoonbatch.job;
-
-import org.springframework.batch.core.Job;
-import org.springframework.batch.core.Step;
-import org.springframework.batch.core.job.builder.JobBuilder;
-import org.springframework.batch.core.repository.JobRepository;
-import org.springframework.batch.core.step.builder.StepBuilder;
-import org.springframework.batch.repeat.RepeatStatus;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.transaction.PlatformTransactionManager;
-
-@Configuration
-public class TestJobConfig {
-
-    @Bean
-    public Job testJob(JobRepository jobRepository, Step testStep) {
-        return new JobBuilder("testJob", jobRepository)
-            .start(testStep)
-            .build();
-    }
-
-    @Bean
-    public Step testStep(JobRepository jobRepository, PlatformTransactionManager transactionManager) {
-        return new StepBuilder("testStep", jobRepository)
-            .tasklet(
-                (contribution, chunkContext) -> {
-                    System.out.println("Test Step executed successfully");
-                    return RepeatStatus.FINISHED;
-                },
-                transactionManager
-            )
-            .build();
-    }
-}
