+++ zipsoonbatch/src/main/java/com/zipsoon/zipsoonbatch/job/reader/NaverClient.java
@@ -27,7 +27,7 @@ public class NaverLandClient {
         maxAttempts = 3,
         backoff = @Backoff(delay = 2000)
     )
-    public NaverArticleResponseDto getArticles(String cortarNo, int page) {
+    public NaverResponseDto getArticles(String cortarNo, int page) {
         HttpHeaders headers = createHeaders();
         HttpEntity<?> requestEntity = new HttpEntity<>(headers);
 
@@ -39,11 +39,11 @@ public class NaverLandClient {
             .build()
             .toUriString();
 
-        ResponseEntity<NaverArticleResponseDto> response = restTemplate.exchange(
+        ResponseEntity<NaverResponseDto> response = restTemplate.exchange(
             url,
             HttpMethod.GET,
             requestEntity,
-            NaverArticleResponseDto.class
+            NaverResponseDto.class
         );
 
         if (response.getStatusCode() != HttpStatus.OK || response.getBody() == null) {
