+++ zipsoonbatch/src/main/java/com/zipsoon/zipsoonbatch/job/reader/NaverResponseDto.java
