+++ zipsoonbatch/src/test/java/com/zipsoon/zipsoonbatch/job/NaverClientTest.java
@@ -2,14 +2,13 @@ package com.zipsoon.zipsoonbatch.job;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.zipsoon.zipsoonbatch.job.reader.NaverArticleResponseDto;
+import com.zipsoon.zipsoonbatch.job.reader.NaverResponseDto;
 import com.zipsoon.zipsoonbatch.job.reader.NaverLandClient;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.HttpStatus;
 import org.springframework.test.context.ActiveProfiles;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -33,13 +32,13 @@ class NaverLandClientTest {
         int page = 1;
 
         // when
-        NaverArticleResponseDto response = naverLandClient.getArticles(cortarNo, page);
+        NaverResponseDto response = naverLandClient.getArticles(cortarNo, page);
 
         // then
         assertThat(response).isNotNull();
         assertThat(response.getArticleList()).isNotEmpty();
 
-        NaverArticleResponseDto.ArticleDto firstArticle = response.getArticleList()[0];
+        NaverResponseDto.ArticleDto firstArticle = response.getArticleList()[0];
         assertThat(firstArticle.getArticleNo()).isNotNull();
         assertThat(firstArticle.getLatitude()).isNotNull();
         assertThat(firstArticle.getLongitude()).isNotNull();
