+++ zipsoonbatch/src/test/java/com/zipsoon/zipsoonbatch/job/PropertyJobTest.java
@@ -18,7 +18,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 @SpringBatchTest
 @SpringBootTest
 @ActiveProfiles("test")
-class PropertyCollectionJobConfigTest {
+class PropertyJobConfigTest {
 
     @Autowired
     private JobLauncherTestUtils jobLauncherTestUtils;
